fix(core): split path helpers into renderer-safe module
ReconcileModal imported `baseNameFor` from the `@contexture/core` barrel,
which transitively pulled `pipeline.ts`'s top-level `node:crypto` import
into the renderer bundle and broke the desktop build.

Move the pure path helpers (no Node deps) into `packages/core/src/paths.ts`,
expose them via a `@contexture/core/paths` subpath, re-export them from
`pipeline.ts` for back-compat, and update the renderer import.
